Large cleanup of Links
